Breakout anisotropic properties into a new base class.
