+++ zephyr/backend/meta.py
@@ -236,3 +236,42 @@ class BaseModelDependent(AttributeMapper):
         else:
             return np.array([lind / self.nx, np.mod(lind, self.nx)]).T
 
+
+class BaseAnisotropic(BaseModelDependent):
+
+    initMap = {
+    #   Argument        Required    Rename as ...   Store as type
+        'theta':        (False,     '_theta',       np.float64),
+        'eps':          (False,     '_eps',         np.float64),
+        'delta':        (False,     '_delta',       np.float64),
+    }
+
+    @property
+    def theta(self):
+        if getattr(self, '_theta', None) is None:
+            self._theta = np.zeros((self.nz, self.nx))
+            
+        if isinstance(self._theta, np.ndarray):
+            return self._theta
+        else:
+            return self._theta * np.ones((self.nz, self.nx), dtype=np.float64)
+
+    @property
+    def eps(self):
+        if getattr(self, '_eps', None) is None:
+            self._eps = np.zeros((self.nz, self.nx))
+            
+        if isinstance(self._eps, np.ndarray):
+            return self._eps
+        else:
+            return self._eps * np.ones((self.nz, self.nx), dtype=np.float64)
+
+    @property
+    def delta(self):
+        if getattr(self, '_delta', None) is None:
+            self._delta = np.zeros((self.nz, self.nx))
+            
+        if isinstance(self._delta, np.ndarray):
+            return self._delta
+        else:
+            return self._delta * np.ones((self.nz, self.nx), dtype=np.float64)
