+++ zephyr/backend/source.py
@@ -1,5 +1,5 @@
 
-from .meta import BaseModelDependent
+from .meta import BaseModelDependent, BaseAnisotropic
 import warnings
 import numpy as np
 import scipy.sparse as sp
@@ -100,7 +100,7 @@ class SparseKaiserSource(SimpleSource):
         10: 14.18,
     }
     
-    def modifyGrid(Zi, Xi):
+    def modifyGrid(self, Zi, Xi):
         
         return Zi, Xi
 
@@ -264,45 +264,8 @@ class KaiserSource(SparseKaiserSource):
         return q.toarray()
 
 
-class AnisotropicSparseKaiserSource(SparseKaiserSource):
+class AnisotropicSparseKaiserSource(SparseKaiserSource, BaseAnisotropic):
     
-    initMap = {
-    #   Argument        Required    Rename as ...   Store as type
-        'theta':        (False,     '_theta',       np.float64),
-        'eps':          (False,     '_eps',         np.float64),
-        'delta':        (False,     '_delta',       np.float64),
-    }
-
-    @property
-    def theta(self):
-        if getattr(self, '_theta', None) is None:
-            self._theta = np.zeros((self.nz, self.nx))
-            
-        if isinstance(self._theta, np.ndarray):
-            return self._theta
-        else:
-            return self._theta * np.ones((self.nz, self.nx), dtype=np.float64)
-
-    @property
-    def eps(self):
-        if getattr(self, '_eps', None) is None:
-            self._eps = np.zeros((self.nz, self.nx))
-            
-        if isinstance(self._eps, np.ndarray):
-            return self._eps
-        else:
-            return self._eps * np.ones((self.nz, self.nx), dtype=np.float64)
-
-    @property
-    def delta(self):
-        if getattr(self, '_delta', None) is None:
-            self._delta = np.zeros((self.nz, self.nx))
-            
-        if isinstance(self._delta, np.ndarray):
-            return self._delta
-        else:
-            return self._delta * np.ones((self.nz, self.nx), dtype=np.float64)
-    
-    def modifyGrid(Zi, Xi):
+    def modifyGrid(self, Zi, Xi):
         
         return Zi, Xi
